# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Video Integration Runtime",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/artifacts/{artifact_id}": {
      "get": {
        "tags": [
          "T05 versions"
        ],
        "summary": "Get Artifact",
        "operationId": "getArtifact",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ArtifactState_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/artifacts/{artifact_id}/revisions": {
      "get": {
        "tags": [
          "T05 versions"
        ],
        "summary": "List Revisions",
        "operationId": "listRevisions",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Artifact Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RevisionPage_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "T05 versions"
        ],
        "summary": "Create Revision",
        "operationId": "createRevision",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Artifact Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_CreateRevision_"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/artifacts/{artifact_id}/adoption-preview": {
      "post": {
        "tags": [
          "T05 versions"
        ],
        "summary": "Preview Adoption",
        "operationId": "previewAdoption",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Artifact Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewAdoption"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Impact_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/artifacts/{artifact_id}/adoptions": {
      "post": {
        "tags": [
          "T05 versions"
        ],
        "summary": "Adopt Revision",
        "operationId": "adoptRevision",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Artifact Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_AdoptRevision_"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/artifacts/{artifact_id}/confirmations": {
      "post": {
        "tags": [
          "T05 versions"
        ],
        "summary": "Confirm Revision",
        "operationId": "confirmRevision",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Artifact Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_ConfirmRevision_"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/adoptions/{adoption_id}/undo": {
      "post": {
        "tags": [
          "T05 versions"
        ],
        "summary": "Undo Adoption",
        "operationId": "undoAdoption",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "adoption_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Adoption Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_UndoAdoption_"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/local-checks": {
      "post": {
        "tags": [
          "T05 versions"
        ],
        "summary": "Run Local Checks",
        "operationId": "runLocalChecks",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command_RunLocalChecks_"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MutationReceipt_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/check-reports/{check_id}": {
      "get": {
        "tags": [
          "T05 versions"
        ],
        "summary": "Get Check Report",
        "operationId": "getCheckReport",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "check_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "format": "uuid",
              "title": "Check Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_CheckReport_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdoptRevision": {
        "properties": {
          "previewId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "toRevisionId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "confirm": {
            "type": "boolean",
            "title": "Confirm"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "previewId",
          "toRevisionId",
          "confirm"
        ],
        "title": "AdoptRevision"
      },
...
      "ArtifactState": {
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/ArtifactKind"
          },
          "adoptedRevisionId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "confirmedRevisionId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "needsUpdate": {
            "type": "boolean",
            "title": "Needsupdate"
          },
          "latestAdoptionId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Uuid"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "adoptedRevisionId",
          "confirmedRevisionId",
          "needsUpdate",
          "latestAdoptionId"
        ],
        "title": "ArtifactState"
      },
...
      "Command_AdoptRevision_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/AdoptRevision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[AdoptRevision]"
      },
...
      "Command_ConfirmRevision_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/ConfirmRevision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[ConfirmRevision]"
      },
...
      "Command_CreateRevision_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/CreateRevision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[CreateRevision]"
      },
...
      "Command_RunLocalChecks_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/RunLocalChecks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[RunLocalChecks]"
      },
...
      "Command_UndoAdoption_": {
        "properties": {
          "clientOperationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "payload": {
            "$ref": "#/components/schemas/UndoAdoption"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clientOperationId",
          "expectedRevision",
          "payload"
        ],
        "title": "Command[UndoAdoption]"
      },
...
      "CreateRevision": {
        "properties": {
          "draftId": {
            "$ref": "#/components/schemas/Uuid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "draftId"
        ],
        "title": "CreateRevision"
      },
...
      "Envelope_ArtifactState_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/ArtifactState"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[ArtifactState]"
      },
...
      "Envelope_CheckReport_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/CheckReport"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[CheckReport]"
      },
...
      "Envelope_Impact_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/Impact"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[Impact]"
      },
...
      "Envelope_MutationReceipt_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/MutationReceipt"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[MutationReceipt]"
      },
...
      "Envelope_RevisionPage_": {
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "data": {
            "$ref": "#/components/schemas/RevisionPage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "data"
        ],
        "title": "Envelope[RevisionPage]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Impact": {
        "properties": {
          "previewId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "artifactId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "fromRevisionId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "toRevisionId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "affectedArtifactIds": {
            "items": {
              "$ref": "#/components/schemas/Uuid"
            },
            "type": "array",
            "maxItems": 10000,
            "title": "Affectedartifactids"
          },
          "affectedScopes": {
            "items": {
              "$ref": "#/components/schemas/SemanticScope"
            },
            "type": "array",
            "maxItems": 9,
            "title": "Affectedscopes"
          },
          "estimatedExtraMicroCny": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SafeInteger"
              },
              {
                "type": "null"
              }
            ]
          },
          "requiredChecks": {
            "items": {
              "$ref": "#/components/schemas/RuleId"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Requiredchecks"
          },
          "expiresAt": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Expiresat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "previewId",
          "artifactId",
          "fromRevisionId",
          "toRevisionId",
          "affectedArtifactIds",
          "affectedScopes",
          "estimatedExtraMicroCny",
          "requiredChecks",
          "expiresAt"
        ],
        "title": "Impact"
      },
...
      "MutationReceipt": {
        "properties": {
          "operationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "committedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          },
          "resourceId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "state": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "State"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operationId",
          "committedRevision",
          "resourceId",
          "state"
        ],
        "title": "MutationReceipt"
      },
...
      "PreviewAdoption": {
        "properties": {
          "toRevisionId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "expectedRevision": {
            "$ref": "#/components/schemas/SafeInteger"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "toRevisionId",
          "expectedRevision"
        ],
        "title": "PreviewAdoption"
      },
...
      "RevisionPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Revision"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Items"
          },
          "nextCursor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Uuid"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "nextCursor"
        ],
        "title": "RevisionPage"
      },
      "RuleId": {
        "type": "string",
        "maxLength": 100,
        "minLength": 1
      },
...
      "SafeInteger": {
        "type": "integer",
        "ge": 0,
        "le": 9007199254740991
      },
...
      "UndoAdoption": {
        "properties": {
          "adoptionId": {
            "$ref": "#/components/schemas/Uuid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "adoptionId"
        ],
        "title": "UndoAdoption"
      },
      "Uuid": {
        "type": "string",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
        "format": "uuid"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "RuntimeBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "32-byte random token, base64url without padding"
      }
    }
  },
  "security": [
    {
      "RuntimeBearer": []
    }
  ]
}